21 tests passing

+++ src/main/java/com/example/repository/MessageRepository.java
@@ -1,11 +1,14 @@
 package com.example.repository;
 import com.example.entity.Message;
 
+import java.util.Optional;
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * JPARepository that will be used to preform persistence operations on Message objects
  */
-public interface MessageRepository extends JpaRepository <Message, Long> {
-    // Message findById(int message_id);
+public interface MessageRepository extends JpaRepository <Message, Integer> {
+    // Optional<Message> findById(Long messageId);
 }
